requirements.txt
amqp==5.0.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==5.0.0
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
Flask==1.0.2
importlib-metadata==2.0.0
isort==5.6.3
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
prompt-toolkit==3.0.7
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.2
toml==0.10.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
